refactor: StatusFor into portfolio-api
